User CRUD permission check added

+++ app/Http/Controllers/Admin/UserController.php
@@ -8,7 +8,7 @@ use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Hash;
-use App\Enums\{RoleEnum, UserStatusEnum};
+use App\Enums\{PermissionEnum, RoleEnum, UserStatusEnum};
 use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 
 class UserController extends Controller
@@ -18,13 +18,21 @@ class UserController extends Controller
      */
     public function index(Request $request)
     {
+        if (auth('web')->user()->cannot(PermissionEnum::USER_READ)) {
+            return redirect()->to(route('dashboard'))
+                ->with([
+                    'success' => false,
+                    'message' => 'You do not have permission to access this page.',
+                ]);
+        }
+
         $search = $request->input('search');
 
         return Inertia::render('admin/User', [
             'users' => (fn () => User::query()
                 ->when($search, function ($query) use ($search): void {
-                    $query->where(function (EloquentBuilder $builder) use ($search): void {
-                        $builder->whereLike('first_name', "%{$search}%")
+                    $query->where(function (EloquentBuilder $eloquentBuilder) use ($search): void {
+                        $eloquentBuilder->whereLike('first_name', "%{$search}%")
                             ->orWhereLike('last_name', "%{$search}%")
                             ->orWhereLike('email', "%{$search}%");
                     });
@@ -38,6 +46,14 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        if (auth('web')->user()->cannot(PermissionEnum::USER_CREATE)) {
+            return redirect()->to(route('dashboard'))
+                ->with([
+                    'success' => false,
+                    'message' => 'You do not have permission to access this page.',
+                ]);
+        }
+
         $validated = $request->validate([
             'first_name' => 'required|string|max:255',
             'last_name'  => 'required|string|max:255',
@@ -58,6 +74,7 @@ class UserController extends Controller
         $user->assignRole(RoleEnum::tryFrom($validated['role']));
 
         return back()->with([
+            'success' => true,
             'message' => 'User created successfully',
         ]);
     }
@@ -67,6 +84,14 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
+        if (auth('web')->user()->cannot(PermissionEnum::USER_UPDATE)) {
+            return redirect()->to(route('dashboard'))
+                ->with([
+                    'success' => false,
+                    'message' => 'You do not have permission to access this page.',
+                ]);
+        }
+
         $validated = $request->validate([
             'first_name' => 'required|string|max:255',
             'last_name'  => 'required|string|max:255',
@@ -96,6 +121,7 @@ class UserController extends Controller
         $user->syncRoles(RoleEnum::tryFrom($validated['role']));
 
         return back()->with([
+            'success' => true,
             'message' => 'User updated successfully',
         ]);
     }
@@ -105,9 +131,18 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if (auth('web')->user()->cannot(PermissionEnum::USER_DELETE)) {
+            return redirect()->to(route('dashboard'))
+                ->with([
+                    'success' => false,
+                    'message' => 'You do not have permission to access this page.',
+                ]);
+        }
+
         $user->delete();
 
         return redirect()->to(route('admin.user.index'))->with([
+            'success' => true,
             'message' => 'User deleted successfully',
         ]);
     }
